split ts components
